Move ::Storage module into Yast::YStorage
* Avoid polluting ::Storage
* Yast::Storage exists yet so a new name was chosen.
