+++ test/Makefile.am
@@ -5,7 +5,7 @@
 TESTS = \
   storage_snapper_configure_snapper_test.rb \
   storage_get_disk_partition.rb \
-  storage/snapshots_finish_test.rb
+  ystorage/snapshots_finish_test.rb
 
 TEST_EXTENSIONS = .rb
 RB_LOG_COMPILER = rspec
